### 2022-10-01 #4812 @editorialbot
[PRE REVIEW]: elapid: Species distribution modeling tools for Python
**Submitting author:** @earth-chris (<a href="http://orcid.org/0000-0001-7392-4368">Christopher Anderson</a>)
**Repository:** https://github.com/earth-chris/elapid
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v0.3.14
