### 2022-10-03 36b17f3
adding missing DOIs
