+++ docs/paper/paper.bib
@@ -8,6 +8,7 @@
   number={4},
   pages={427--433},
   year={1917},
+  doi={10.2307/4072271},
   publisher={JSTOR}
 }
 
@@ -19,6 +20,7 @@
   number={supplement\_2},
   pages={19729--19736},
   year={2009},
+  doi={10.1073/pnas.0901639106},
   publisher={National Acad Sciences}
 }
 
@@ -30,6 +32,7 @@
   number={1},
   pages={1--9},
   year={2012},
+  doi={10.1111/j.1365-2699.2011.02637.x},
   publisher={Wiley Online Library}
 }
 
@@ -41,6 +44,7 @@
   number={7},
   pages={887--893},
   year={2017},
+  doi={10.1111/ecog.03049},
   publisher={Wiley Online Library}
 }
 
@@ -52,6 +56,7 @@
   number={2},
   pages={161--175},
   year={2008},
+  doi={10.1111/j.0906-7590.2008.5203.x},
   publisher={Wiley Online Library}
 }
 
@@ -63,6 +68,7 @@
   number={1},
   pages={1--9},
   year={2014},
+  doi={10.1111/ddi.12144},
   publisher={Wiley Online Library}
 }
 
@@ -83,6 +89,7 @@
   number={10},
   pages={1058--1069},
   year={2013},
+  doi={10.1111/j.1600-0587.2013.07872.x},
   publisher={Wiley Online Library}
 }
 
@@ -94,6 +101,7 @@
   number={1},
   pages={43--57},
   year={2011},
+  doi={10.1111/j.1472-4642.2010.00725.x},
   publisher={Wiley Online Library}
 }
 
@@ -105,6 +113,7 @@
   number={4},
   pages={1917},
   year={2013},
+  doi={10.1214/13-aoas667},
   publisher={NIH Public Access}
 }
 
@@ -116,6 +125,7 @@
   number={2},
   pages={245--256},
   year={2018},
+  doi={10.1111/geb.12684},
   publisher={Wiley Online Library}
 }
 
@@ -127,6 +137,7 @@
   number={2},
   pages={327--338},
   year={2012},
+  doi={10.1111/j.2041-210x.2011.00172.x},
   publisher={Wiley Online Library}
 }
 
@@ -138,6 +149,7 @@
   number={1},
   pages={128--141},
   year={2022},
+  doi={10.1111/ddi.13442},
   publisher={Wiley Online Library}
 }
 
