### 2022-10-03 db1c562
added predict_proba method
