+++ elapid/models.py
@@ -4,12 +4,19 @@ from warnings import warn
 
 import numpy as np
 import pandas as pd
-from sklearn.base import BaseEstimator
+from sklearn.base import BaseEstimator, ClassifierMixin
 from sklearn.exceptions import NotFittedError
 from sklearn.linear_model import LogisticRegression
 
-from elapid import features as _features
 from elapid.config import MaxentConfig, NicheEnvelopeConfig
+from elapid.features import (
+    CategoricalTransformer,
+    FeaturesMixin,
+    MaxentFeatureTransformer,
+    compute_lambdas,
+    compute_regularization,
+    compute_weights,
+)
 from elapid.types import ArrayLike, Number, validate_feature_types
 from elapid.utils import NCPUS, make_band_labels
 
@@ -26,7 +33,7 @@ except ModuleNotFoundError:
     FORCE_SKLEARN = True
 
 
-class MaxentModel(BaseEstimator):
+class MaxentModel(BaseEstimator, ClassifierMixin):
     """Model estimator for Maxent-style species distribution models."""
 
     # passed to __init__
@@ -166,7 +173,7 @@ class MaxentModel(BaseEstimator):
                 x = self.preprocessor.fit_transform(x)
 
         # fit the feature transformer
-        self.transformer = _features.MaxentFeatureTransformer(
+        self.transformer = MaxentFeatureTransformer(
             feature_types=self.feature_types,
             clamp=self.clamp,
             n_hinge_features=self.n_hinge_features,
@@ -178,7 +185,7 @@ class MaxentModel(BaseEstimator):
         # compute class weights
         if self.class_weights is not None:
             pbr = len(y) / y.sum() if self.class_weights == "balanced" else self.class_weights
-            class_weight = _features.compute_weights(y, pbr=pbr)
+            class_weight = compute_weights(y, pbr=pbr)
 
             # scale the sample weight
             if sample_weight is None:
@@ -196,7 +203,7 @@ class MaxentModel(BaseEstimator):
         # model fitting with glmnet
         else:
             # set feature regularization parameters
-            self.regularization_ = _features.compute_regularization(
+            self.regularization_ = compute_regularization(
                 y,
                 features,
                 feature_labels=feature_labels,
@@ -208,7 +215,7 @@ class MaxentModel(BaseEstimator):
             )
 
             # get model lambda scores to initialize the glm
-            self.lambdas_ = _features.compute_lambdas(y, sample_weight, self.regularization_, n_lambdas=self.n_lambdas)
+            self.lambdas_ = compute_lambdas(y, sample_weight, self.regularization_, n_lambdas=self.n_lambdas)
 
             # model fitting
             self.initialize_glmnet_model(lambdas=self.lambdas_)
@@ -236,6 +243,8 @@ class MaxentModel(BaseEstimator):
 
         # the distance from f(z) is the relative entropy of f1(z) WRT f(z)
         self.entropy_ = maxent_entropy(raw)
+
+        return self
 
     def predict(self, x: ArrayLike, transform: str = "cloglog") -> ArrayLike:
         """Apply a model to a set of covariates or features. Requires that a model has been fit.
@@ -356,7 +365,7 @@ class MaxentModel(BaseEstimator):
         )
 
 
-class NicheEnvelopeModel(BaseEstimator):
+class NicheEnvelopeModel(BaseEstimator, ClassifierMixin, FeaturesMixin):
     """Model estimator for niche envelope-style models."""
 
     percentile_range: Tuple[float, float] = None
@@ -379,71 +388,6 @@ class NicheEnvelopeModel(BaseEstimator):
                 covariates at all y==1 locations.
         """
         self.percentile_range = percentile_range
[…]
