### 2022-10-03 9c2f03a
added TransformerMixin methods
