+++ elapid/features.py
@@ -4,7 +4,7 @@ from typing import Any, List, Tuple, Union
 
 import numpy as np
 import pandas as pd
-from sklearn.base import BaseEstimator
+from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.preprocessing import MinMaxScaler, OneHotEncoder, QuantileTransformer
 
 from elapid.config import MaxentConfig, RegularizationConfig
@@ -96,7 +96,7 @@ class LinearTransformer(MinMaxScaler):
         super().__init__(clip=clamp, feature_range=feature_range)
 
 
-class QuadraticTransformer(BaseEstimator):
+class QuadraticTransformer(BaseEstimator, TransformerMixin):
     """Applies quadtratic feature transformations and rescales features from 0-1."""
 
     clamp: bool = None
@@ -110,19 +110,22 @@ class QuadraticTransformer(BaseEstimator):
     ):
         self.clamp = clamp
         self.feature_range = feature_range
-        self.estimator = MinMaxScaler(clip=self.clamp, feature_range=self.feature_range)
 
-    def fit(self, x: ArrayLike) -> None:
+    def fit(self, x: ArrayLike, y: None = None) -> "QuadraticTransformer":
         """Compute the minimum and maximum for scaling.
 
         Args:
             x: array-like of shape (n_samples, n_features)
                 The data used to compute the per-feature minimum and maximum
                 used for later scaling along the features axis.
+
         Returns:
-            None. Updates the transformer with feature fitting parameters.
+            self. Returns the transformer with fitted parameters.
         """
+        self.estimator = MinMaxScaler(clip=self.clamp, feature_range=self.feature_range)
         self.estimator.fit(np.array(x) ** 2)
+
+        return self
 
     def transform(self, x: ArrayLike) -> np.ndarray:
         """Scale covariates according to the feature range.
@@ -162,7 +165,7 @@ class QuadraticTransformer(BaseEstimator):
         return self.estimator.inverse_transform(np.array(x)) ** 0.5
 
 
-class ProductTransformer(BaseEstimator):
+class ProductTransformer(BaseEstimator, TransformerMixin):
     """Computes the column-wise product of an array of input features, rescaling from 0-1."""
 
     clamp: bool = None
@@ -176,19 +179,22 @@ class ProductTransformer(BaseEstimator):
     ):
         self.clamp = clamp
         self.feature_range = feature_range
-        self.estimator = MinMaxScaler(clip=self.clamp, feature_range=self.feature_range)
 
-    def fit(self, x: ArrayLike) -> None:
+    def fit(self, x: ArrayLike) -> "ProductTransformer":
         """Compute the minimum and maximum for scaling.
 
         Args:
             x: array-like of shape (n_samples, n_features)
                 The data used to compute the per-feature minimum and maximum
                 used for later scaling along the features axis.
+
         Returns:
-            None. Updates the transformer with feature fitting parameters.
+            self. Returns the transformer with fitted parameters.
         """
+        self.estimator = MinMaxScaler(clip=self.clamp, feature_range=self.feature_range)
         self.estimator.fit(column_product(np.array(x)))
+
+        return self
 
     def transform(self, x: ArrayLike) -> np.ndarray:
         """Scale covariates according to the feature range.
@@ -216,7 +222,7 @@ class ProductTransformer(BaseEstimator):
         return self.transform(x)
 
 
-class ThresholdTransformer(BaseEstimator):
+class ThresholdTransformer(BaseEstimator, TransformerMixin):
     """Applies binary thresholds to each covariate based on n evenly-spaced
     thresholds across it's min/max range."""
 
@@ -228,20 +234,23 @@ class ThresholdTransformer(BaseEstimator):
     def __init__(self, n_thresholds: int = MaxentConfig.n_threshold_features):
[…]
