### 2022-10-03 3b98cfe
removing parameter validation from init
