### 2022-10-03 abda5c3
tested predict_proba() methods
