### 2022-10-05 2f6b0f6
overlay and transform arguments handled at class instead of at predict level
