+++ elapid/config.py
@@ -29,6 +29,7 @@ class MaxentConfig:
 
     # species prevalence scalar
     tau: float = 0.5
+    transform: str = "cloglog"
 
     # model convergence tolerance threshold
     tolerance: float = 1e-7
@@ -53,3 +54,4 @@ class RegularizationConfig:
 
 class NicheEnvelopeConfig:
     percentile_range: Tuple[float, float] = [2.5, 97.5]
+    overlay: str = "average"
