+++ elapid/models.py
@@ -70,6 +70,7 @@ class MaxentModel(BaseEstimator, ClassifierMixin):
         self,
         feature_types: Union[list, str] = MaxentConfig.feature_types,
         tau: float = MaxentConfig.tau,
+        transform: float = MaxentConfig.transform,
         clamp: bool = MaxentConfig.clamp,
         scorer: str = MaxentConfig.scorer,
         beta_multiplier: float = MaxentConfig.beta_multiplier,
@@ -92,6 +93,8 @@ class MaxentModel(BaseEstimator, ClassifierMixin):
             feature_types: maxent feature types to fit. must be in string "lqphta" or
                 list ["linear", "quadratic", "product", "hinge", "threshold", "auto"]
             tau: maxent prevalence value for scaling logistic output
+            transform: maxent model transformation type. select from
+                ["raw", "logistic", "cloglog"].
             clamp: set features to min/max range from training during prediction
             scorer: sklearn scoring function for model training
             beta_multiplier: scaler for all regularization parameters.
@@ -118,6 +121,7 @@ class MaxentModel(BaseEstimator, ClassifierMixin):
         """
         self.feature_types = feature_types
         self.tau = tau
+        self.transform = transform
         self.clamp = clamp
         self.scorer = scorer
         self.beta_multiplier = beta_multiplier
@@ -237,7 +241,10 @@ class MaxentModel(BaseEstimator, ClassifierMixin):
         self.initialized_ = True
 
         # apply maxent-specific transformations
-        raw = self.predict(x[y == 0], transform="raw")
+        class_transform = self.get_params()["transform"]
+        self.set_params(transform="raw")
+        raw = self.predict(x[y == 0])
+        self.set_params(transform=class_transform)
 
         # alpha is a normalizing constant that ensures that f1(z) integrates (sums) to 1
         self.alpha_ = maxent_alpha(raw)
@@ -247,13 +254,11 @@ class MaxentModel(BaseEstimator, ClassifierMixin):
 
         return self
 
-    def predict(self, x: ArrayLike, transform: str = "cloglog") -> ArrayLike:
+    def predict(self, x: ArrayLike) -> ArrayLike:
         """Apply a model to a set of covariates or features. Requires that a model has been fit.
 
         Args:
             x: array-like of shape (n_samples, n_features) with covariate data
-            transform: maxent model transformation type. select from
-                ["raw", "logistic", "cloglog"].
 
         Returns:
             predictions: array-like of shape (n_samples,) with model predictions
@@ -269,27 +274,25 @@ class MaxentModel(BaseEstimator, ClassifierMixin):
         engma = np.matmul(features, self.beta_scores_) + self.alpha_
 
         # scale based on the transform type
-        if transform == "raw":
+        if self.transform == "raw":
             return maxent_raw_transform(engma)
 
-        elif transform == "logistic":
+        elif self.transform == "logistic":
             return maxent_logistic_transform(engma, self.entropy_, self.tau)
 
-        elif transform == "cloglog":
+        elif self.transform == "cloglog":
             return maxent_cloglog_transform(engma, self.entropy_)
 
-    def predict_proba(self, x: ArrayLike, transform: str = "cloglog") -> ArrayLike:
+    def predict_proba(self, x: ArrayLike) -> ArrayLike:
         """Compute prediction probability scores for the 0/1 classes.
 
         Args:
             x: array-like of shape (n_samples, n_features) with covariate data
-            transform: maxent model transformation type. select from
-                ["raw", "logistic", "cloglog"].
 
         Returns:
             predictions: array-like of shape (n_samples, 2) with model predictions
         """
[…]
