### 2022-10-05 9b8b1f5
added EnsembleModel class
