### 2022-10-05 712f352
added anaconda version badge
