+++ docs/index.md
@@ -7,8 +7,9 @@
 </p>
 
 ![GitHub](https://img.shields.io/github/license/earth-chris/elapid)
-![PyPI](https://img.shields.io/pypi/v/elapid)
-![GitHub workflow wtatus](https://img.shields.io/github/workflow/status/earth-chris/elapid/docs)
+![PyPI version](https://img.shields.io/pypi/v/elapid)
+![Anaconda version](https://anaconda.org/conda-forge/elapid/badges/version.svg)
+![GitHub workflow status](https://img.shields.io/github/workflow/status/earth-chris/elapid/docs)
 ![PyPI downloads](https://img.shields.io/pypi/dm/elapid)
 ![GitHub last commit](https://img.shields.io/github/last-commit/earth-chris/elapid)
 [![JOSS manuscript status](https://joss.theoj.org/papers/ac415a024261efb3b397a1bad6f9cde6/status.svg)](https://earth-chris.github.io/elapid/paper/draft-paper.pdf)
