### 2022-10-05 32d1c6d
fixing links to install page
