### 2022-10-05 574e80f
adding conda install
