### 2022-10-05 6450a12
adding progress bar to point annotation
