+++ requirements.txt
@@ -3,6 +3,7 @@ geopandas>=0.7.0
 glmnet;python_version<'3.9'
 numpy>=1.18
 pandas>=1.0.3
+pyproj>=3.0
 rasterio>=1.2.1
 rtree>=0.9
 scikit-learn>=0.22.2
