### 2022-10-11 #4812 @danielskatz
commands to @editorialbot need to be the first thing in a comment 🙂
