### 2022-10-12 abb445a
removing intro paragraph
