+++ docs/paper/paper.md
@@ -21,11 +21,9 @@ bibliography: paper.bib
 
 # Summary
 
-Biogeographers use species occurrence records and environmental data to understand the current, past, and future spatial distributions of biodiversity. `elapid` is a contemporary SDM package built on best practices from the past, and aspires to help build the next generation of biodiversity modeling tools.
-
 Species distribution modeling (SDM) is based on the Grinellean niche concept: the environmental conditions that allow individuals of a species to survive and reproduce will constrain the distributions of those species over space and time [@Grinnell:1917; @Wiens:2009]. The inputs to these models are typically spatially-explicit species occurrence records and a series of environmental covariates, which might include information on climate, topography, land cover or hydrology [@Booth:2014]. While many modeling methods have been developed to quantify and map these species-environment interactions, few software systems include both a) the appropriate statistical modeling routines and b) support for handling the full suite of geospatial analysis required to prepare data to fit, apply, and summarize these models.
 
-`elapid` is both a geospatial analysis and a species distribution modeling package. It provides an interface between vector and raster data for selecting random point samples, annotating point locations with coincident raster data, and summarizing raster values inside a polygon with zonal statistics. It provides a series of covariate transformation routines for increasing feature dimensionality, quantifying interaction terms and normalizing unit scales. It provides a Python implementation of the popular Maxent SDM [@Phillips:2017] using infinitely weighted logistic regression [@Fithian:2013]. It also includes a standard Niche Envelope Model [@Nix:1986], both of which were written to match the software design patterns of modern machine learning packages like `sklearn` [@sklearn]. It also allows users to add spatial context to any model by providing methods for spatially splitting train/test data and computing geographically-explicit sample weights.
+`elapid` is both a geospatial analysis and a species distribution modeling package. It provides an interface between vector and raster data for selecting random point samples, annotating point locations with coincident raster data, and summarizing raster values inside a polygon with zonal statistics. It provides a series of covariate transformation routines for increasing feature dimensionality, quantifying interaction terms and normalizing unit scales. It provides a Python implementation of the popular Maxent SDM [@Phillips:2017] using infinitely weighted logistic regression [@Fithian:2013]. It also includes a standard Niche Envelope Model [@Nix:1986], both of which were written to match the software design patterns of modern machine learning packages like `sklearn` [@sklearn]. It also allows users to add spatial context to any model by providing methods for spatially splitting train/test data and computing geographically-explicit sample weights. `elapid` was designed as a contemporary SDM package, built on best practices from the past and aspiring to support the next generation of biodiversity modeling workflows.
 
 # Statement of need
 
