### 2022-10-24 51c4953
v0 buffered leave-one-out implementation
