### 2022-10-24 8fb071a
0.3.18 adds the BufferedLeaveOneOut class
