### 2022-10-25 c2158c4
added class_label support for buffered leave-one-out
