+++ elapid/train_test_split.py
@@ -7,6 +7,7 @@ import numpy as np
 from shapely.geometry import box
 from sklearn.cluster import KMeans
 from sklearn.model_selection import BaseCrossValidator
+from sklearn.utils.validation import _num_samples
 
 from elapid.geo import nearest_point_distance
 from elapid.types import Vector
@@ -73,9 +74,25 @@ class GeographicKFold(BaseCrossValidator):
     """Compute geographically-clustered train/test folds using KMeans clustering"""
 
     def __init__(self, n_splits: int = 4):
+        """Cluster x/y points into separate cross-validation folds.
+
+        Args:
+            n_splits: Number of geographic clusters to split the data into.
+        """
         self.n_splits = n_splits
 
-    def split(self, points: Vector) -> Tuple[gpd.GeoDataFrame, gpd.GeoDataFrame]:
+    def _iter_test_indices(self, points: Vector, y: None = None, groups: None = None):
+        """Generate indices for test data samples."""
+        kmeans = KMeans(n_clusters=self.n_splits)
+        xy = np.array(list(zip(points.geometry.x, points.geometry.y)))
+        kmeans.fit(xy)
+        clusters = kmeans.predict(xy)
+        indices = np.arange(len(xy))
+        for cluster in range(self.n_splits):
+            test = clusters == cluster
+            yield indices[test]
+
+    def split(self, points: Vector) -> Tuple[np.ndarray, np.ndarray]:
         """Split point data into geographically-clustered train/test folds and
             return their array indices.
 
@@ -88,51 +105,41 @@ class GeographicKFold(BaseCrossValidator):
         for train, test in super().split(points):
             yield train, test
 
-    def _iter_test_indices(self, X: Vector, y: None = None, groups: None = None):
-        """The method used by the base class to split train/test data"""
-        kmeans = KMeans(n_clusters=self.n_splits)
-        xy = np.array(list(zip(X.geometry.x, X.geometry.y)))
-        kmeans.fit(xy)
-        clusters = kmeans.predict(xy)
-        indices = np.arange(len(xy))
-        for cluster in range(self.n_splits):
-            test = clusters == cluster
-            yield indices[test]
-
-    def get_n_splits(self, X: None = None, y: None = None, groups: None = None) -> int:
-        """Returns the number of splitting iterations in the cross-validator
-
-        Args:
-            X: ignored, exists for compatibility.
-            y: ignored, exists for compatibility.
-            groups: ignored, exists for compatibility.
+    def get_n_splits(self) -> int:
+        """Return the number of splitting iterations in the cross-validator.
 
         Returns:
-            The number of splitting iterations in the cross-validator.
+            Splitting iteration count.
         """
         return self.n_splits
 
 
 class BufferedLeaveOneOut(BaseCrossValidator):
-    """Leave-one-out CV that excludes nearby training data points."""
+    """Leave-one-out CV that excludes training points within a buffered distance."""
 
     distance: float = None
 
     def __init__(self, distance: float):
-        """Leave-one-out cross-validation strategy.
+        """Buffered leave-one-out cross-validation strategy.
 
         Drops points from the training data based on a buffered distance
-            from the left-out test point.
-
-        Implemented as described in Ploton et al. 2020
+            to the left-out test point(s). Implemented from Ploton et al. 2020,
             https://www.nature.com/articles/s41467-020-18321-y
+
+        Args:
+            distance: drop training data points within this distance of test data.
         """
         self.distance = distance
 
-    def _group_idxs(self, points, groups, count: bool = False) -> List[int]:
-        """Get the test indices for group train/test splits."""
-        unique = points[groups].unique()
+    def _group_idxs(
[…]
