### 2022-11-01 #4812 @chrisborges
Hi Openjournals/Joss-Reviews,

I am available to review Dr. Higino's submission.

On Tue, Nov 1, 2022 at 8:33 AM Juan Escamilla Molgora <
***@***.***> wrote:

> Hi Gracielle,I might have some time off but I’m not sure. What would be
> the expected timeframe at the moment?Cheers,JuanEl 1 nov 2022, a las 3:07,
> Gracielle Higino ***@***.***> escribió:﻿
> @molgor - would you be willing to review this submission? 😄
>
> —Reply to this email directly, view it on GitHub, or unsubscribe.You are
> receiving this because you were mentioned.Message ID: ***@***.***>
>
