### 2022-11-07 #4812 @graciellehigino
> Hi Gracielle,I might have some time off but I’m not sure. What would be the expected timeframe at the moment?Cheers,JuanEl 1 nov 2022, a las 3:07, Gracielle Higino ***@***.***> escribió:﻿ @molgor - would you be willing to review this submission? 😄 —Reply to this email directly, view it on GitHub, or unsubscribe.You are receiving this because you were mentioned.Message ID: ***@***.***>

Hi @molgor ! We'd have about 3 weeks for review once the review issue is open. If you agree, we can open it right away. I'll try to get a third reviewer in the meantime.
