### 2022-11-11 1ffa4d2
set hinge, threshold attributes to expeced sklearn values
