### 2022-11-11 35a5816
converted class attributes to instance attributes
