### 2022-11-12 fa92b16
added buffered leave-one-out class_label docs
