+++ requirements.txt
@@ -1,6 +1,5 @@
 descartes>=1.1.0
 geopandas>=0.7.0
-glmnet;python_version<'3.9'
 numpy>=1.18
 pandas>=1.0.3
 pyproj>=3.0
