### 2022-11-12 ce6ea42
simplified install section
