### 2022-11-14 #4812 @graciellehigino
Thank you very much for agreeing to review, @chrisborges and @gabrieldansereau! I really appreciate it. @molgor I'd very much like to add you to the review team, let me know if you are interested as soon as possible 😃

You can find reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), together with all the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).

I will start the review issue where you will find a copy of the manuscript, a link to the software repository, and a checklist. During the review, you can check to see if the software repository satisfies the items on the checklist. If not, we encourage reviewers to file issues against the submitted repository’s issue tracker. When you have completed your review, please leave a comment in the review issue saying so.

Happy reviewing!
