### 2022-11-14 #4930 @editorialbot
[REVIEW]: elapid: Species distribution modeling tools for Python
**Submitting author:** @earth-chris (<a href="http://orcid.org/0000-0001-7392-4368">Christopher Anderson</a>)
**Repository:** https://github.com/earth-chris/elapid
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.1
