+++ docs/paper/paper.md
@@ -31,19 +31,19 @@ Species occurrence data—georeferenced point locations where a species has been
 
 `elapid` builds on a suite of well-known statistical modeling tools commonly used by biogeographers, extending them to add novel features, to work with cloud-hosted data, and to save and share models. It provides methods for managing the full lifecyle of modeling data: generating background point data, extracting raster values for each point (i.e. point annotation), splitting train/test data, fitting models, and applying predictions to rasters. It provides a very high degree of control for model design, which is important for several reasons.
 
-First is to provide simple and flexible methods for working with spatial data. Point data are managed as `GeoSeries` and `GeoDataFrame` objects [@geopandas], which can be easily merged and split using traditional indexing method as well as with geograhic methods. They can also be reprojected on-the-fly. `elapid` reads and writes raster data with `rasterio`, which provides a similarly convenient set of methods for indexing and reading point locations from rasters [@rasterio]. These features are wrapped to handle many of the routine tasks and gotchas of working with geospatial data. It doesn't require data to be rigorously pre-processed so that all rasters are perfectly aligned, nor does it require that all datasets are in matching projections. `elapid` can extract pixel-level raster data from datasets at different resolutions, from multi-band files, and harmonize projections on-the-fly, for both model fitting and for inference.
+First is to provide simple and flexible methods for working with spatial data. Point data are managed as `GeoSeries` and `GeoDataFrame` objects [@geopandas], which can be easily merged and split using traditional indexing method as well as with geographic methods. They can also be reprojected on-the-fly. `elapid` reads and writes raster data with `rasterio`, which provides a similarly convenient set of methods for indexing and reading point locations from rasters [@rasterio]. These features are wrapped to handle many of the routine tasks and gotchas of working with geospatial data. It doesn't require data to be rigorously pre-processed so that all rasters are perfectly aligned, nor does it require that all datasets are in matching projections. `elapid` can extract pixel-level raster data from datasets at different resolutions, from multi-band files, and harmonize projections on-the-fly, for both model fitting and for inference.
 
-Another advantage of `elapid`'s flexible design is that it can be used to extend traditional species distribution models in ways that are difficult to implement in other software systems. Working with multi-temporal data, for example—fitting SDMs to occurrence records and environmental data from multiple time periods—is also supported. Each time period's occurrence data can be annotated using the coincident environmental data. Random background samples can likewise be generated for each time period, which ensures the background represents a broad distribution of conditions across the full temporal extent. These presence and background samples then concatenated into a single `GeoDataFrame` for model fitting. Fitted models can be applied to multi-temporal environmental data to map changes in habitat suitability over time, and can also be saved and restored later for future inference.
+Another advantage of `elapid`'s flexible design is that it can be used to extend traditional species distribution models in ways that are difficult to implement in other software systems. Working with multi-temporal data, for example—fitting SDMs to occurrence records and environmental data from multiple time periods—is also supported. Each time period's occurrence data can be annotated using the coincident environmental data. Random background samples can likewise be generated for each time period, which ensures the background represents a broad distribution of conditions across the full temporal extent. These presence and background samples are then concatenated into a single `GeoDataFrame` for model fitting. Fitted models can be applied to multi-temporal environmental data to map changes in habitat suitability over time, and can also be saved and restored later for future inference.
 
 # Why Maxent still matters
 
 The main scientific contribution of `elapid` is extending and modifying the Maxent SDM, a model and software system as popular as it is maligned [@Phillips:2008; @Fourcade:2018]. First published in 2006, Maxent remains relevant because it's a presence-only model designed to work with the kinds of species occurrence data data that have proliferated lately.
 
[…]
