### 2022-11-15 earth-chris/elapid#74 @gabrieldansereau
[JOSS review]: Sample data missing at installation
Hi @earth-chris! I'm opening a few issues here for my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4930).These will be mostly about bugs or issues when trying to reproduce the examples in your documentation.

I'm having trouble loading the sample data with `elapid.load_sample_data()`. The files appear to be missing from my local installation (as you can see in the output below). I doubled-checked at the path given in the error message (`/home/gabriel/.local/lib/python3.10/site-packages/elapid/`) and I simply do not have a `data` folder or the `bradypus.csv.gz` file.

I managed to debug it by downloading the data from this repo and moving it to a `data` folder in the right path. After that `elapid.load_sample_data()` works as expected.

I installed elapid using `pip install elapid` as explained in the documentation. I'm using elapid v0.3.18 and Python 3.10.6 on Ubuntu 22.04.

Here is the detailed output:
```
Python 3.10.6 (main, Nov  2 2022, 18:53:38) [GCC 11.3.0] on linux
>>> import elapid
>>> x, y = elapid.load_sample_data()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/gabriel/.local/lib/python3.10/site-packages/elapid/utils.py", line 54, in load_sample_data
    df = pd.read_csv(file_path, compression="gzip").astype("int64")
  File "/home/gabriel/.local/lib/python3.10/site-packages/pandas/util/_decorators.py", line 211, in wrapper
    return func(*args, **kwargs)
  File "/home/gabriel/.local/lib/python3.10/site-packages/pandas/util/_decorators.py", line 331, in wrapper
    return func(*args, **kwargs)
  File "/home/gabriel/.local/lib/python3.10/site-packages/pandas/io/parsers/readers.py", line 950, in read_csv
    return _read(filepath_or_buffer, kwds)
  File "/home/gabriel/.local/lib/python3.10/site-packages/pandas/io/parsers/readers.py", line 605, in _read
    parser = TextFileReader(filepath_or_buffer, **kwds)
  File "/home/gabriel/.local/lib/python3.10/site-packages/pandas/io/parsers/readers.py", line 1442, in __init__
    self._engine = self._make_engine(f, self.engine)
  File "/home/gabriel/.local/lib/python3.10/site-packages/pandas/io/parsers/readers.py", line 1735, in _make_engine
    self.handles = get_handle(
  File "/home/gabriel/.local/lib/python3.10/site-packages/pandas/io/common.py", line 750, in get_handle
    handle = gzip.GzipFile(  # type: ignore[assignment]
  File "/usr/lib/python3.10/gzip.py", line 174, in __init__
    fileobj = self.myfileobj = builtins.open(filename, mode or 'rb')
FileNotFoundError: [Errno 2] No such file or directory: '/home/gabriel/.local/lib/python3.10/site-packages/elapid/data/bradypus.csv.gz'
```
