### 2022-11-15 earth-chris/elapid#74 @earth-chris
👋🏻 hi @gabrieldansereau! thanks for opening all these issues - I'm grateful for your attention. I'll work to resolve many of them shortly. Many should be low effort, others will require a bit more time, but they're all clear and straightforward to address.
