### 2022-11-15 #4930 @earth-chris
Thank you @gabrieldansereau, I appreciate your feedback. These are critical issues to address, which I will do shortly. I'll report back to this thread again once complete.

@chrisborges, I suspect these issues would limit your reviewing experience, and I might recommend waiting until I address them before providing your review, which should reduce your review burden.
