### 2023-02-03 #4930 @earth-chris
Hi @graciellehigino (and reviewers!). My apologies for the delay. I won't bore you with the details, but approximately everything happened all at once in my life. I appreciate your patience as I return to some semblance of normal.

I plan to address the most critical issues this weekend (providing / debugging sample data to demonstrate reproducible results). If I'm lucky I can get the docs updated, too. I'll be sure to ping this thread once I do so. But if this weekend proves too ambitious, I will prioritize unblocking everyone by the end of the coming week.
