+++ mkdocs.yml
@@ -78,6 +78,11 @@ extra_css:
     - https://earth-chris.github.io/shared/mkdocs.css
     #- css/colors.css
 
+extra_javascript:
+  - javascript/mathjax.js
+  - https://polyfill.io/v3/polyfill.min.js?features=es6
+  - https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js
+
 # pandoc (https://pandoc.org/MANUAL.html)
 markdown_extensions:
     - admonition
@@ -86,7 +91,8 @@ markdown_extensions:
         guess_lang: false
     - def_list
     - footnotes
-    - pymdownx.arithmatex
+    - pymdownx.arithmatex:
+        generic: true
     - pymdownx.betterem
     - pymdownx.caret:
         insert: false
