### 2023-02-04 9dac4c1
Add mathjax (#83)
* added mathjax features

* formatting primary maxent formulation
