+++ elapid/geo.py
@@ -305,6 +305,7 @@ def annotate(
     raster_paths: Union[str, list],
     labels: list = None,
     drop_na: bool = True,
+    quiet: bool = False,
 ) -> gpd.GeoDataFrame:
     """Read raster values for each point in a vector and append as new columns.
 
@@ -316,6 +317,7 @@ def annotate(
         labels: band name labels. number of labels should match the
             total number of bands across all raster_paths.
         drop_na: drop all records with no-data values.
+        quiet: silence progress bar output.
 
     Returns:
         GeoDataFrame annotated with the pixel values from each raster
@@ -331,6 +333,7 @@ def annotate(
             raster_paths,
             labels=labels,
             drop_na=drop_na,
+            quiet=quiet,
         )
 
     elif isinstance(points, gpd.GeoDataFrame) or isinstance(points, pd.DataFrame):
@@ -339,13 +342,14 @@ def annotate(
             raster_paths,
             labels=labels,
             drop_na=drop_na,
+            quiet=quiet,
         )
 
         # append annotations to the input dataframe
         gdf = pd.concat([points, gdf.drop(["geometry"], axis=1, errors="ignore")], axis=1)
 
     elif os.path.isfile(points):
-        gdf = annotate_vector(points, raster_paths, labels=labels, drop_na=drop_na)
+        gdf = annotate_vector(points, raster_paths, labels=labels, drop_na=drop_na, quiet=quiet)
 
     else:
         raise TypeError("points arg must be a valid path, GeoDataFrame, or GeoSeries")
@@ -358,6 +362,7 @@ def annotate_vector(
     raster_paths: list,
     labels: list = None,
     drop_na: bool = True,
+    quiet: bool = False,
 ) -> gpd.GeoDataFrame:
     """Reads and stores pixel values from rasters using a point-format vector file.
 
@@ -366,6 +371,7 @@ def annotate_vector(
         raster_paths: raster paths to extract pixel values from
         labels: band name labels. should match the total number of bands across all raster_paths
         drop_na: drop all records with no-data values
+        quiet: silence progress bar output.
 
     Returns:
         gdf: GeoDataFrame annotated with the pixel values from each raster
@@ -381,7 +387,12 @@ def annotate_vector(
 
 
 def annotate_geoseries(
-    points: gpd.GeoSeries, raster_paths: list, labels: list = None, drop_na: bool = True, dtype: str = None
+    points: gpd.GeoSeries,
+    raster_paths: list,
+    labels: list = None,
+    drop_na: bool = True,
+    dtype: str = None,
+    quiet: bool = False,
 ) -> gpd.GeoDataFrame:
     """Reads and stores pixel values from rasters using point locations.
 
@@ -391,6 +402,7 @@ def annotate_geoseries(
         labels: band labels. must match the total number of bands for all raster_paths.
         drop_na: drop records with no-data values.
         dtype: output column data type. uses the first raster's dtype by default.
+        quiet: silence progress bar output.
 
     Returns:
         gdf: GeoDataFrame annotated with the pixel values from each raster
@@ -408,7 +420,9 @@ def annotate_geoseries(
     nodata_flag = False
 
     # annotate each point with the pixel values for each raster
-    for raster_idx, raster_path in tqdm(enumerate(raster_paths), desc="Raster", total=n_rasters, **tqdm_opts):
+    for raster_idx, raster_path in tqdm(
+        enumerate(raster_paths), desc="Raster", total=n_rasters, disable=quiet, **tqdm_opts
+    ):
         with rio.open(raster_path, "r") as src:
 
             # reproject points to match raster and convert to a dataframe
@@ -425,7 +439,14 @@ def annotate_geoseries(
             # read each pixel value
[…]
