### 2023-02-05 17c5807
moved tqdm_opts to utils
