### 2023-02-05 04efb7b
set drop_geometry default to False
