+++ tests/test_features.py
@@ -1,7 +1,7 @@
 from elapid import features
 from elapid.utils import load_sample_data
 
-x, y = load_sample_data()
+x, y = load_sample_data(name="ariolimax", drop_geometry=True)
 nrows, ncols = x.shape
 
 
