### 2023-02-05 2363b24
add ariolimax dataset (#85)
* fixing default xy crs

* added ariolimax buttoni sample data

* added more general sample data tests

* added download_sample_data function

* moved tqdm_opts to utils

* set drop_geometry default to False
