### 2023-02-06 e97918c
stack_geometries -> stack_geodataframes
