+++ tests/test_geo.py
@@ -32,20 +32,20 @@ def test_xy_to_geoseries():
     assert geoseries.y[0] == lat
 
 
-def test_stack_geometries():
+def test_stack_geodataframes():
     pts = gpd.read_file(points)
     n_pts = len(pts)
-    stack = geo.stack_geometries(pts, pts)
+    stack = geo.stack_geodataframes(pts, pts)
     assert len(stack) == 2 * n_pts
 
-    stack = geo.stack_geometries(pts, pts, add_class_label=True)
+    stack = geo.stack_geodataframes(pts, pts.copy(), add_class_label=True)
     assert stack["class"].sum() == n_pts
 
     new_crs = "EPSG:7844"
-    stack = geo.stack_geometries(pts, pts.to_crs(new_crs))
+    stack = geo.stack_geodataframes(pts, pts.to_crs(new_crs))
     assert geo.crs_match(stack.crs, pts.crs)
 
-    stack = geo.stack_geometries(pts, pts.to_crs(new_crs), target_crs="background")
+    stack = geo.stack_geodataframes(pts, pts.to_crs(new_crs), target_crs="background")
     assert geo.crs_match(stack.crs, new_crs)
 
 
