### 2023-02-08 dea6567
updated python api basics
