### 2023-02-08 3c53f6b
updated drop_na logic
