### 2023-02-08 d1780a8
add notebook docs (#86)
* stack_geometries -> stack_geodataframes

* adding geo data notebook

* init

* copy edits, bugfixes

* removing markdown examples

* updated python api basics

* updated drop_na logic

* updated version number, plot

* 0.3.19 includes annotate bugfixes, stack_geodataframes update, adds the quiet keyword for tqdm

* clead up output cells

* updated order, title of example docs
