+++ elapid/geo.py
@@ -56,7 +56,7 @@ def xy_to_geoseries(
     return gs
 
 
-def stack_geometries(
+def stack_geodataframes(
     presence: Vector, background: Vector, add_class_label: bool = False, target_crs: str = "presence"
 ) -> gpd.GeoDataFrame:
     """Concatenate geometries from two GeoSeries/GeoDataFrames.
@@ -71,29 +71,35 @@ def stack_geometries(
     Returns:
         merged GeoDataFrame with all geometries projected to the same crs.
     """
-    pgeo = presence.geometry
-    bgeo = background.geometry
+    validate_gpd(presence)
+    validate_gpd(background)
 
+    # cast to geodataframes
+    if isinstance(presence, gpd.GeoSeries):
+        presence = presence.to_frame("geometry")
+    if isinstance(background, gpd.GeoSeries):
+        background = background.to_frame("geometry")
+
+    # handle projection mismatch
     crs = presence.crs
     if not crs_match(presence.crs, background.crs):
         if target_crs.lower() == "presence":
-            bgeo = bgeo.to_crs(crs)
+            background.to_crs(crs, inplace=True)
         elif target_crs.lower() == "background":
             crs = background.crs
-            pgeo = pgeo.to_crs(crs)
+            presence.to_crs(crs, inplace=True)
         else:
             raise NameError(f"Unrecognized target_crs option: {target_crs}")
 
-    attributes = {}
     if add_class_label:
-        npresence = len(pgeo)
-        nbackground = len(bgeo)
-        classes = np.zeros(npresence + nbackground, dtype="uint8")
-        classes[:npresence] = 1
-        attributes["class"] = classes
+        presence["class"] = 1
+        background["class"] = 0
 
-    geometry = pd.concat((pgeo, bgeo), axis=0, ignore_index=True)
-    gdf = gpd.GeoDataFrame(attributes, geometry=geometry, crs=crs)
+    matching = [col for col in presence.columns if col in background.columns]
+    assert len(matching) > 0, "no matching columns found between data frames"
+
+    merged = pd.concat((presence[matching], background[matching]), axis=0, ignore_index=True)
+    gdf = gpd.GeoDataFrame(merged, crs=crs)
 
     return gdf
 
@@ -328,6 +334,7 @@ def annotate(
 
     # read raster values based on the points dtype
     if isinstance(points, gpd.GeoSeries):
+        points = points.reset_index(drop=True)
         gdf = annotate_geoseries(
             points,
             raster_paths,
@@ -337,6 +344,7 @@ def annotate(
         )
 
     elif isinstance(points, gpd.GeoDataFrame) or isinstance(points, pd.DataFrame):
+        points = points.reset_index(drop=True)
         gdf = annotate_geoseries(
             points.geometry,
             raster_paths,
@@ -353,6 +361,13 @@ def annotate(
 
     else:
         raise TypeError("points arg must be a valid path, GeoDataFrame, or GeoSeries")
+
+    if drop_na:
+        try:
+            valid = gdf["valid"] == 1
+            gdf = gdf[valid].drop(columns="valid").dropna().reset_index(drop=True)
+        except KeyError:
+            pass
 
     return gdf
 
@@ -383,6 +398,7 @@ def annotate_vector(
     gdf = gpd.read_file(vector_path)
     raster_df = annotate_geoseries(gdf.geometry, raster_paths, labels, drop_na)
     gdf = pd.concat([gdf, raster_df.drop(["geometry"], axis=1, errors="ignore")], axis=1)
+
     return gdf
 
 
@@ -393,7 +409,7 @@ def annotate_geoseries(
     drop_na: bool = True,
     dtype: str = None,
     quiet: bool = False,
-) -> gpd.GeoDataFrame:
[…]
