+++ elapid/__version__.py
@@ -1,1 +1,1 @@
-"0.3.18"
+"0.3.19"
