### 2023-02-08 #4930 @earth-chris
hello again @graciellehigino, @gabrieldansereau and @chrisborges. I've made a series of key updates in response to Gabriel's review, focused on ensuring reproducible results and clear documentation.

In short, I've added new sample data with all the fields required to run the suite of spatial and statistical analysis provided by `elapid`. And I've updated the documentation to be a jupyter notebook that can be run end-to-end (or the snippets could be copied and pasted to run).

I'll take another pass at the manuscript to flesh out the related software/research references, but as of now I believe you all should be unblocked from moving forward with your reviews. Thanks again for your patience.
