### 2023-02-09 0cf4761
added option to automatically calculate the distance to all neighbors
