### 2023-02-09 0afccae
setting new default for distance_weights
