+++ elapid/geo.py
@@ -888,14 +888,15 @@ def nearest_point_distance(
     return distance.mean(axis=1)
 
 
-def distance_weights(points: Vector, n_neighbors: int = 1, center: str = "median", cpu_count: int = -1) -> np.ndarray:
+def distance_weights(points: Vector, n_neighbors: int = -1, center: str = "median", cpu_count: int = -1) -> np.ndarray:
     """Compute sample weights based on the distance between points.
 
     Assigns higher scores to isolated points, lower scores to clustered points.
 
     Args:
         points: point-format GeoSeries or GeoDataFrame
-        n_neighbors: compute the average distance to the nearest n_neighbors
+        n_neighbors: compute weights based on average distance to the nearest n_neighbors
+            set to -1 to compute the distance to all neighbors.
         center: rescale the weights to center the mean or median of the array on 1
             accepts either 'mean' or 'median' as input.
             pass None to ignore.
