### 2023-02-20 earth-chris/elapid#88 @aFewThings
Citing elapid library
Hi, I want to cite the elapid library in my paper.
Is your library plan to be published as a paper, or is there any proper way to cite elapid?
