### 2023-02-28 4df336a
calculate distance weights from all neighbors (#87)
* added option to automatically calculate the distance to all neighbors

* setting new default for distance_weights

* modified default distance weights section
