### 2023-02-28 earth-chris/elapid#88 @earth-chris
hi @aFewThings, thanks for reaching out. `elapid` has been submitted to JOSS and you can find the review [here](https://github.com/openjournals/joss-reviews/issues/4930). I'm not sure when that will wrap up. Do you have a sense for when you're planning to submit (or, when you'll need a proper DOI)?
