### 2023-02-28 earth-chris/elapid#88 @earth-chris
In the short-term, I've added a file with the basic software information you can use to cite the package. Just click the link on the right-hand side of the repository to get the formatted citation:

![image](https://user-images.githubusercontent.com/10216182/221777681-a01d0aad-6af9-4879-b741-795274f3a716.png)

I'll avoid generating a DOI at this time as I'd prefer just wait until I get one from the JOSS submission.
