### 2023-03-01 #4930 @gabrieldansereau
Hi @earth-chris! Thank you for fixing the reproducibility issues. I was able to reproduce all the examples in the documentation. The geospatial and Maxent vignettes are very detailed and give a great idea on how to use `elapid`. It more than meets the JOSS requirements in my opinion.

The last comment to address is the State of the field criteria for the Software paper. I'm reposting from my earlier comment here. I know `maxnet` and others are well discussed in the documentation, but this is a criteria for the Software paper specifically, so I do think there should be some mention in there as well.

> The software paper is very clear and well-written. All criteria are met, but I think there could be a few more elements to meet the "State of the field" criteria of comparing the software with commonly-used packages. For example, I realized that the documentation refers to `maxnet`, the R version of Maxent, but the software paper does not. I think it would be relevant to mention it in the software paper as well.
> This also made me wonder if you should mention R packages as `dismo` or `sdm`, which offer similar species distribution modelling features. Is `elapid` the first package to provide similar integrated species distribution modelling features in Python? If so, this would be a great selling point for it.
