+++ elapid/__version__.py
@@ -1,1 +1,1 @@
-"0.3.19"
+"0.3.20"
