+++ docs/paper/paper.bib
@@ -153,6 +153,48 @@
   publisher={Wiley Online Library}
 }
 
+@article{Kass:2021,
+  author = {Kass, Jamie M. and Muscarella, Robert and Galante, Peter J. and Bohl, Corentin L. and Pinilla-Buitrago, Gonzalo E. and Boria, Robert A. and Soley-Guardia, Mariano and Anderson, Robert P.},
+  title = {ENMeval 2.0: Redesigned for customizable and reproducible modeling of species’ niches and distributions},
+  journal = {Methods in Ecology and Evolution},
+  volume = {12},
+  number = {9},
+  pages = {1602-1608},
+  keywords = {cross-validation, ecological niche model, metadata, model evaluation, model tuning, software, spatial, species distribution model},
+  doi = {https://doi.org/10.1111/2041-210X.13628},
+  url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.13628},
+  eprint = {https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.13628},
+  abstract = {Abstract Quantitative evaluations to optimize complexity have become standard for avoiding overfitting of ecological niche models (ENMs) that estimate species’ potential geographic distributions. ENMeval was the first R package to make such evaluations (often termed model tuning) widely accessible for the Maxent algorithm. It also provided multiple methods for partitioning occurrence data and reported various performance metrics. Requests by users, recent developments in the field, and needs for software compatibility led to a major redesign and expansion. We additionally conducted a literature review to investigate trends in ENMeval use (2015–2019). ENMeval 2.0 has a new object-oriented structure for adding other algorithms, enables customizing algorithmic settings and performance metrics, generates extensive metadata, implements a null-model approach to quantify significance and effect sizes, and includes features to increase the breadth of analyses and visualizations. In our literature review, we found insufficient reporting of model performance and parameterization, heavy reliance on model selection with AICc and low utilization of spatial cross-validation; we explain how ENMeval 2.0 can help address these issues. This redesigned and expanded version can promote progress in the field and improve the information available for decision-making.},
+  year = {2021}
+}
+
+@article{Brown:2014,
+  author = {Brown, Jason L.},
+  title = {SDMtoolbox: a python-based GIS toolkit for landscape genetic, biogeographic and species distribution model analyses},
+  journal = {Methods in Ecology and Evolution},
+  volume = {5},
+  number = {7},
+  pages = {694-700},
+  keywords = {ArcGIS, geographic information systems, least-cost corridors, corrected weighted endemism, ecological niche models, Maxent bias files, spatially rarefy occurrences, spatial jackknifing},
+  doi = {https://doi.org/10.1111/2041-210X.12200},
+  url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.12200},
+  eprint = {https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.12200},
[…]
