+++ docs/paper/paper.md
@@ -33,17 +33,15 @@ Species occurrence data—georeferenced point locations where a species has been
 
 First is to provide simple and flexible methods for working with spatial data. Point data are managed as `GeoSeries` and `GeoDataFrame` objects [@geopandas], which can be easily merged and split using traditional indexing method as well as with geographic methods. They can also be reprojected on-the-fly. `elapid` reads and writes raster data with `rasterio`, which provides a similarly convenient set of methods for indexing and reading point locations from rasters [@rasterio]. These features are wrapped to handle many of the routine tasks and gotchas of working with geospatial data. It doesn't require data to be rigorously pre-processed so that all rasters are perfectly aligned, nor does it require that all datasets are in matching projections. `elapid` can extract pixel-level raster data from datasets at different resolutions, from multi-band files, and harmonize projections on-the-fly, for both model fitting and for inference.
 
-Another advantage of `elapid`'s flexible design is that it can be used to extend traditional species distribution models in ways that are difficult to implement in other software systems. Working with multi-temporal data, for example—fitting SDMs to occurrence records and environmental data from multiple time periods—is also supported. Each time period's occurrence data can be annotated using the coincident environmental data. Random background samples can likewise be generated for each time period, which ensures the background represents a broad distribution of conditions across the full temporal extent. These presence and background samples are then concatenated into a single `GeoDataFrame` for model fitting. Fitted models can be applied to multi-temporal environmental data to map changes in habitat suitability over time, and can also be saved and restored later for future inference.
+Another advantage of `elapid`'s flexible design is that it can be used to extend traditional species distribution models in ways that are difficult to implement in other software systems. For example, working with multi-temporal data—fitting SDMs to occurrence records and environmental data from multiple time periods—is supported. Each time period's occurrence data can be annotated using the coincident environmental data. Random background samples can be generated for each time period, ensuring the background represents a broad distribution of conditions across the full temporal extent. These presence and background samples are then concatenated into a single `GeoDataFrame` for model fitting. Fitted models can be applied to multi-temporal environmental data to map changes in habitat suitability over time, and can also be saved and restored later for future inference.
 
-Compare to other software. maxnet, ENMeval as comparable, which calls maxnet and dismo under the hood. elapid less focused on providing automated tools for model selection (e.g. through aic), more on expanding the range of available spatial xval methods (they use checkerboard, too). Design choices to allow for more flexible feature engineering (sklearn api design), making predictions on non-raster files (i.e. tabular data in pandas), and flexibility for handling resampling based on the relevant scale (like annotate tool from jetz lab).
-
-ENMEval: https://doi.org/10.1111/2041-210X.13628
+`elapid` is one among several open source species distribution modeling packages. The R package `ENMeval` is a good direct comparison [@Kass:2021]. `ENMeval` provides a series of tools for model fitting, model selection and cross-validation, making calls under the hood to `maxnet` and `dismo` [@Phillips:2017]. `elapid` implements similar methods for spatial cross-validation, builds on the rich feature transformation tools implemented in `maxnet`, and employs similar model fitting techniques. `elapid` provides additional tools to simplify working with geospatial datasets, and provides additional novel cross-validation methods like geographic *k*-fold and buffered leave-one-out strategies [@Ploton:2020]. It is also one of the first open source species distribution modeling packages in Python, and it does not include any proprietary software dependencies [@Brown:2014].
 
 # Why Maxent still matters
 
 The main scientific contribution of `elapid` is extending and modifying the Maxent SDM, a model and software system as popular as it is maligned [@Phillips:2008; @Fourcade:2018]. First published in 2006, Maxent remains relevant because it's a presence-only model designed to work with the kinds of species occurrence data data that have proliferated lately.
 
[…]
