### 2023-03-01 #4930 @earth-chris
thanks @gabrieldansereau! I've added a new paragraph pointing out a comparable package (i think `ENMeval` is pretty close in a few ways), mentioned the similar lineage to the R `maxnet` package, added an additional cross-validation reference, and highlighted the novelty of this python package (the only other established python SDM package is effectively an ArcGIS extentsion).
