### 2023-03-12 #4930 @chrisborges
Hi @earth-chris and @graciellehigino. I apologize truly for my delay in reviewing. First thing, I want to acknowledge @gabrieldansereau's amazing work in reviewing this paper, he has given amazing feedback and has surely made my work easier.

_Software paper_
I have no qualms with the software paper - I think it is clear, well-written and to-the-point.

_Documentation_
The documentation is also well-written and clear. However, on your readme ˜Why use elapid?” section, the examples link is broken (https://github.com/earth-chris/elapid/blob/main/examples/geo). I was able to find the folder with notebook examples in your repository. But this link should be fixed as it will probably be the first click for users.

_Installation and reproducibility
Installation took a while, but worked as expected.

WorkingWithGeospatialData notebook is great. I was able to reproduce almost everything and enjoyed the through explanation. Unfortunately, I was not able to reproduce the histogram graph.

some things that threw errors: _TypeError: zonal_stats() got an unexpected keyword argument 'quiet'_

In both "WorkingWithGeospatialData" and "A simple maxent model" I was unable to run the merged command:
merged = ela.stack_geodataframes(presence, background, add_class_label=True)
Issue:  _AttributeError: module 'elapid' has no attribute 'stack_geodataframes'_

Thus, due to **this error** specifically, I was unable to finish the examples and do not confirm the functional claims of the software at the moment.

Once this issues are resolved, I would love to resume with the examples.
