### 2023-03-13 #4930 @earth-chris
hi @chrisborges! thanks for reviewing this.

I've fixed the broken link on the main page - thanks for pointing this out.

Regarding the remaining errors, I suspect you may be using an out-of-date branch. Following Gabriel's input, and a series of other updates, a few new versions of the software have been released since the original submission.

If you've installed from source, please pull the most recent updates from the `main` branch. If you installed it from pip, please try running `pip install --upgrade elapid`.

If you've tried this and are still stuck then please let me know and I'll dig deeper.
