+++ .pre-commit-config.yaml
@@ -3,25 +3,25 @@ default_language_version:
 
 repos:
   - repo: https://github.com/ambv/black
-    rev: 22.8.0
+    rev: 23.1.0
     hooks:
       - id: black
         args: ["elapid", "--line-length=120", "--target-version=py37"]
 
   - repo: https://github.com/pycqa/flake8
-    rev: 5.0.4
+    rev: 6.0.0
     hooks:
       - id: flake8
         args: ["--select=C,E,F,W,B,B950", "--max-line-length=120", "--ignore=E203,E501,W503,F401,F403"]
 
   - repo: https://github.com/timothycrosley/isort
-    rev: 5.10.1
+    rev: 5.12.0
     hooks:
       - id: isort
         args: ["-l 120", "--profile", "black", "."]
 
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.3.0
+    rev: v4.4.0
     hooks:
       - id: check-yaml
         exclude: recipe/meta.yaml
