### 2023-03-14 earth-chris/elapid#92 @earth-chris
package install challenges from `conda-forge`
@chrisborges reported the following issues when installing from conda:

> WorkingWithGeospatialData notebook is great. I was able to reproduce almost everything and enjoyed the through explanation. Unfortunately, I was not able to reproduce the histogram graph.
> some things that threw errors: TypeError: zonal_stats() got an unexpected keyword argument 'quiet'
> In both "WorkingWithGeospatialData" and "A simple maxent model" I was unable to run the merged command:
> merged = ela.stack_geodataframes(presence, background, add_class_label=True)
> Issue: AttributeError: module 'elapid' has no attribute 'stack_geodataframes'

These appear to be version issues that should be resolved in `0.3.20`. But they persisted even after running `conda update elapid`.

Unfortunately I have not been able to reproduce these errors. I was able to run each demo notebook end-to-end after installing elapid in a clean conda environment.

To diagnose, @chrisborges, would you be able to run the following?

```python
import elapid
print(elapid.__version__)
```

If the version isn't `0.3.20` then we'll want to make sure you get the latest version. To do that, you could try the following:

- `conda install elapid=0.3.20`
- or install in a clean environment using [these instructions](https://earth-chris.github.io/elapid/install/#resolving-geospatial-dependencies)

If you do have the most recent version and you're still experiencing issues then I'll have to go back to the drawing board.
