### 2023-03-17 earth-chris/elapid#93 @earth-chris
Support partial dependence plots and feature importance scores
The Java implementation of Maxent includes tools for creating partial dependence plots and quantifying permutation importance. These haven't been ported to `elapid` yet but should be.

This can be implemented using native sklearn tools, including [permutation_importance](https://scikit-learn.org/stable/modules/permutation_importance.html) and [partial_dependence](https://scikit-learn.org/stable/modules/generated/sklearn.inspection.partial_dependence.html#sklearn.inspection.partial_dependence).

Maybe we can add these as class methods for all elapid models?
