### 2023-03-20 earth-chris/elapid#92 @chrisborges
Hey Chris,

So I tried running the version command and got an error
`NameError: name 'elapid' is not defined`

I later tried to install using the elapid=0.3.20 command and got this:
```
PackagesNotFoundError: The following packages are not available from current channels:
  - elapid=0.3.20
```

I think the problem might be with my OS - which is macOS Ventura. Any thoughts?
