### 2023-03-21 d7939a9
prioritizing pip over conda install
