### 2023-03-21 f4f158a
created new mixin class to use auc as default scorer
