### 2023-04-01 #4930 @earth-chris
Hi @graciellehigino!

As far as I'm concerned, all the software works as stated and as intended. I would very much like to conclude this review.

I'm reasonably confident that @chrisborges issues with reproducibility arose as a result of installing an old version of the software that didn't include recent(-ish) updates. I've been providing support to get things installed over the past three weeks, but she has not been available.

I've modified the install instructions to be a bit more clear. I've tested that installing from both pip and conda works on both OSX and Linux. I've run the notebooks used for documentation on three separate machines. All unit and integration tests are set up and running on four versions of python, and on OSX, Linux, and Windows. This seems, to me, well-tested for reproducibility.

At this point I'm just waiting for feedback or approval from @chrisborges. Is there anything else I need to do @graciellehigino?
