### 2023-04-01 #4930 @chrisborges
@graciellehigino Unfortunately I have not been able to test this software and its usage example.

The first time ever I tried to install this package was on March 12, following strictly the install instructions provided in the documentation. If this was an older version then it was not updated on March 12. Since then, I have not been able to get the package working, despite @earth-chris`s continuous support and provided solutions.

It has indeed been a long process, and I have been busy for some of it (we all have, I would say). Perhaps it would be a good idea to bring in a third party to ensure reproducibility, as I haven't been able to.
