### 2023-04-05 #4930 @earth-chris
I'd prefer to avoid waiting to find a third reviewer. The key issue that @chrisborges had was reproducing the first two code blocks from [this example](https://earth-chris.github.io/elapid/examples/MaxentSimpleModel/). My intuition is this is probably an issue on @chrisborges side and not a bug with `elapid` (more on this below).

tl;dr: to wrap this up, @graciellehigino would you be willing to [install](https://earth-chris.github.io/elapid/install/) the package and run those two blocks to confirm they work?

---

Here's a review of the "reproducibility" issue. The error @chrisborges encountered was:

```
AttributeError: module 'elapid' has no attribute 'stack_geodataframes'
```

And as a result she "was unable to finish the examples and [does] not confirm the functional claims of the software at the moment."

But `elapid` does include a `stack_geodataframes()` method. You can see this function [here](https://github.com/earth-chris/elapid/blob/main/elapid/geo.py#L59), which was refactored in [the PR for v0.3.19](https://github.com/earth-chris/elapid/pull/86). It's tested [here](https://github.com/earth-chris/elapid/blob/main/tests/test_geo.py#L35), which is run on CI for [4 python versions](https://github.com/earth-chris/elapid/blob/main/.github/workflows/run-pytest.yml) and [three operating systems](https://github.com/earth-chris/elapid/blob/main/.github/workflows/os-tests.yml). This [documentation page](https://earth-chris.github.io/elapid/examples/MaxentSimpleModel/) is actually a jupyter notebook that I run end to end during each version upgrade, which is inherently validating the functional claims of the software (including that elapid does indeed include a `stack_geodataframes` method). I post all this just as evidence confirming the functional claims of the software.

My suspicion is that, because @chrisborges installed from `conda`, that she likely encountered a version mismatch. `conda` has a complex dependency resolver and it's not always the case that the most recent version will get installed - it depends on what was already installed in the environment she used. I've tried providing support to debug this and verify, but I haven't been able to get any further details ([tracked here](https://github.com/earth-chris/elapid/issues/92)).

---

Another point on reproducibility, and driving my desire to conclude this review, is that `elapid` has hundreds of monthly downloads, and I haven't received any feedback regarding installation issues (save for this [resolved issue](https://github.com/earth-chris/elapid/issues/9)). I receive regular messages from users in the EU, US, and Asia using elapid, I'm being asked about [citation info](https://github.com/earth-chris/elapid/issues/88) and about developing new features ([this request](https://github.com/earth-chris/elapid/issues/93) came from a linkedin message).

I mention this because I would very much like to finish this review, publish version 1.0, and advertise that this package is available for wider use within the biogeography community, since the people who use it quite like it. I've also got a few personal projects that are waiting for this to conclude as well. I hope we can do so soon.
