### 2023-04-05 #4930 @chrisborges
I agree with @earth-chris that the problem I am facing with the package is most likely local and I apologize, but I do not have the time this week to troubleshoot this. I really do not want to delay this anymore either.

It seems `elapid ` is already being used around the world, which puts me at ease concerning reproducibility issues. I also agree with the solution proposed of @graciellehigino running those codes to guarantee they work.

I reiterate I am satisfied with the written portion of this software, and have no pending issues with the work other then not running those 2 lines. @earth-chris has been very supportive and I believe his eagerness to publish the software is completely justifiable.
