### 2023-04-10 #4930 @graciellehigino
Hi all! I tried the code here and all seems to work fine. Some results are a bit different - not in the magnitude of changing conclusions -, but I assume this is normal given the nature of the method. I'm happy to recommend this paper for publication! =D

@earth-chris the next steps for you now are to issue a new tagged release of the software (if changed), archive it, then post the version number and archive DOI in this issue. When this is done, we can generate the final proof of the publication and finalize the review.

I'm tagging @kthyng (EiC) already so we can move a bit faster at this phase. [=

Congratulations, @earth-chris ! 🎉
