### 2023-04-10 3867f01
removing categorical_features kwarg for permutation importance
