### 2023-04-10 d892f46
explicitly handling categorical features under the hood
