+++ tests/test_models.py
@@ -8,6 +8,7 @@ from elapid import models
 from elapid.utils import load_sample_data
 
 x, y = load_sample_data(name="ariolimax", drop_geometry=True)
+xb, yb = load_sample_data("bradypus")
 
 
 def test_MaxentModel_flow():
@@ -197,20 +198,32 @@ def test_EnsembleModel():
 def test_partial_dependence_scores():
     ne = models.NicheEnvelopeModel()
     me = models.MaxentModel()
-    ne.fit(x, y)
-    me.fit(x, y)
 
     # just test that these methods work with each estimator
+    ne.fit(x, y)
+    me.fit(x, y)
     ne.partial_dependence_plot(x)
     me.partial_dependence_plot(x)
+
+    # and with the bradypus data
+    ne.fit(xb, yb)
+    me.fit(xb, yb)
+    ne.partial_dependence_plot(xb, categorical_features=[2])
+    me.partial_dependence_plot(xb)
 
 
 def test_permutation_importance_scores():
     ne = models.NicheEnvelopeModel()
     me = models.MaxentModel()
-    ne.fit(x, y)
-    me.fit(x, y)
 
     # just test that these methods work with each estimator
+    ne.fit(x, y)
+    me.fit(x, y)
     ne.permutation_importance_plot(x, y)
     me.permutation_importance_plot(x, y)
+
+    # and with bradypus
+    ne.fit(xb, yb)
+    me.fit(xb, yb)
+    ne.permutation_importance_plot(xb, yb)
+    me.permutation_importance_plot(xb, yb)
