### 2023-04-10 9a06c83
Add model inspection tools (#95)
* created new mixin class to use auc as default scorer

* 0.3.21 adds new model inspection features

* added permutation importance methods

* setting channel for devl mamba install

* removing pd import

* fleshed out permutation importance plots

* reverting to default maxnet background class weights

* v0 partial dependence plots

* adding partial dependence, permutation importance tests

* removing categorical_features kwarg for permutation importance

* added partial dependence plots

* explicitly handling categorical features under the hood
