+++ elapid/models.py
@@ -1,13 +1,14 @@
 """Classes for training species distribution models."""
 from typing import List, Tuple, Union
-from warnings import warn
 
+import matplotlib.pyplot as plt
 import numpy as np
-import pandas as pd
 from scipy import stats as scistats
-from sklearn.base import BaseEstimator, ClassifierMixin
+from sklearn.base import BaseEstimator
 from sklearn.exceptions import NotFittedError
+from sklearn.inspection import partial_dependence, permutation_importance
 from sklearn.linear_model import LogisticRegression
+from sklearn.metrics import roc_auc_score
 
 from elapid.config import EnsembleConfig, MaxentConfig, NicheEnvelopeConfig
 from elapid.features import (
@@ -18,8 +19,8 @@ from elapid.features import (
     compute_regularization,
     compute_weights,
 )
-from elapid.types import ArrayLike, Number, validate_feature_types
-from elapid.utils import NCPUS, make_band_labels
+from elapid.types import ArrayLike, validate_feature_types
+from elapid.utils import NCPUS, make_band_labels, square_factor
 
 # handle windows systems without functioning gfortran compilers
 FORCE_SKLEARN = False
@@ -30,7 +31,222 @@ except ModuleNotFoundError:
     FORCE_SKLEARN = True
 
 
-class MaxentModel(BaseEstimator, ClassifierMixin):
+class SDMMixin:
+    """Mixin class for SDM classifiers."""
+
+    _estimator_type = "classifier"
+    classes_ = [0, 1]
+
+    def score(self, x: ArrayLike, y: ArrayLike, sample_weight: ArrayLike = None) -> float:
+        """Return the mean AUC score on the given test data and labels.
+
+        Args:
+            x: test samples. array-like of shape (n_samples, n_features).
+            y: presence/absence labels. array-like of shape (n_samples,).
+            sample_weight: array-like of shape (n_samples,)
+
+        Returns:
+            AUC score of `self.predict(x)` w.r.t. `y`.
+        """
+        return roc_auc_score(y, self.predict(x), sample_weight=sample_weight)
+
+    def _more_tags(self):
+        return {"requires_y": True}
+
+    def permutation_importance_scores(
+        self,
+        x: ArrayLike,
+        y: ArrayLike,
+        sample_weight: ArrayLike = None,
+        n_repeats: int = 10,
+        n_jobs: int = -1,
+    ) -> np.ndarray:
+        """Compute a generic feature importance score by modifying feature values
+            and computing the relative change in model performance.
+
+        Permutation importance measures how much a model score decreases when a
+            single feature value is randomly shuffled. This score doesn't reflect
+            the intrinsic predictive value of a feature by itself, but how important
+            feature is for a particular model.
+
+        Args:
+            x: test samples. array-like of shape (n_samples, n_features).
+            y: presence/absence labels. array-like of shape (n_samples,).
+            sample_weight: array-like of shape (n_samples,)
+            n_repeats: number of permutation iterations.
+            n_jobs: number of parallel compute tasks. set to -1 for all cpus.
+
+        Returns:
+            importances: an array of shape (n_features, n_repeats).
+        """
+        pi = permutation_importance(self, x, y, sample_weight=sample_weight, n_jobs=n_jobs, n_repeats=n_repeats)
+
+        return pi.importances
+
+    def permutation_importance_plot(
+        self,
+        x: ArrayLike,
+        y: ArrayLike,
+        sample_weight: ArrayLike = None,
+        n_repeats: int = 10,
+        labels: list = None,
+        **kwargs,
+    ) -> Tuple[plt.Figure, plt.Axes]:
+        """Create a box plot with bootstrapped permutation importance scores for each covariate.
+
+        Permutation importance measures how much a model score decreases when a
[…]
