+++ elapid/__version__.py
@@ -1,1 +1,1 @@
-"0.3.21"
+"1.0.1"
