+++ docs/examples/geo.md
@@ -1,384 +1,0 @@
-```python
-import elapid as ela
-```
-
----
-
-Almost all of the `elapid` point sampling and indexing tools use `geopandas.GeoDataFrame` or `geopandas.GeoSeries` objects. The latter are the format of the `geometry` column for a GeoDataFrame.
-
-```python
-import geopandas as gpd
-
-gdf = gpd.read_file('/path/to/some/vector.shp')
-print(type(gdf.geometry))
-
-> <class 'geopandas.geoseries.GeoSeries'>
-```
-
-Several `elapid` routines return GeoSeries objects (like `ela.sample_raster()`). Others return GeoDataFrames (like `ela.annotate()` and `ela.zonal_stats()`). Be prepared to work with GeoPandas objects.
-
----
-
-## Working with X/Y data
-
-**From CSVs**
-
-Sometimes you don't have a vector of point-format location data. You may instead have a series of x/y locations in a CSV file, for example. You can convert those using the `elapid.xy_to_geoseries()` function:
-
-```python
-import pandas as pd
-import elapid as ela
-
-csv_path = "/home/slug/ariolimax-californicus.csv"
-df = pd.read_csv(csv_path)
-presence = ela.xy_to_geoseries(df.x, df.y, crs="EPSG:32610")
-```
-
-This assumes that the input CSV file has an `x` and a `y` column.
-
-Be sure to specify the projection of your x/y data! The default assumption is lat/lon, which in many cases is not correct.
-
-**From arrays or lists**
-
-You can also convert arbitrary arrays of x/y data to GeoSeries:
-
-```python
-lons = [-122.49, 151.0]
-lats = [37.79, -33.87]
-locations = ela.xy_to_geoseries(lons, lats)
-print(locations)
-
-> 0    POINT (-122.49000 37.79000)
-> 1    POINT (151.00000 -33.87000)
-> dtype: geometry
-```
-
----
-
-## Drawing point location samples
-
-In addition to species occurrence records (where `y = 1`), species distributions models often require a set of random pseudo-absence/background points (`y = 0`). These are a random geographic sampling of where you might expect to find a species across the target landscape.
-
-**From a raster's extent**
-
-You can use `elapid` to create a uniform random geographic sample from unmasked locations within a raster's extent:
-
-```python
-count = 10000 # the number of points to generate
-pseudoabsence_points = ela.sample_raster(raster_path, count)
-```
-
-If you have a large raster that doesn't fit in memory, you can also set `ignore_mask=True` to use the rectangular bounds of the raster to draw samples.
-
-```python
-pseudoabsence_points = ela.sample_raster(raster_path, count, ignore_mask=True)
-```
-
-**From a vector polygon**
-
-Species occurrence records are often collected with some bias (near roads, in protected areas, etc.), so we typically need to be more precise in where we select background points. One way to do this would be to draw samples from that species range map, which you can do with the `sample_vector()` function:
-
-```python
-range_path = "/home/slug/ariolimax-californicus-range.shp"
-pseudoabsence_points = ela.sample_vector(range_path, count)
-```
-
-This currently computes the spatial union of all polygons within a vector, compressing the geometry into a single MultiPolygon object to sample.
-
-If you've already computed a polygon using geopandas, you can pass it instead to `ela.sample_geoseries()`, which is what `sample_vector()` calls under the hood.
-
-**From a bias raster**
-
-You could also pass a raster bias file, where the raster grid cells contain information on the probability of sampling an area:
-
-```python
-# assume covariance between vertebrate and invertebrate banana slugs
-bias_path = "/home/slug/proximity-to-uc-santa-cruz.tif"
-pseudoabsence_points = ela.sample_bias_file(bias_path)
-```
-
[…]
