### 2023-04-10 #4930 @earth-chris
thanks so much @graciellehigino! 🎉

big thanks to @gabrieldansereau and @chrisborges! `elapid` is much higher quality software now thanks to your input. 🐍

version: v1.0.1
DOI: https://doi.org/10.5281/zenodo.7813017

👋🏻 @kthyng (big fan of `cmocean` over here, nice to meet you)
