### 2023-04-11 #4930 @kthyng
I have read through this page now so I am a bit caught up. @earth-chris thanks, glad you like cmocean!

I understand that @chrisborges was not able to run all of the steps required for review and was therefore not able to finalize the review, leaving three checkboxes unmarked. @graciellehigino could you state which steps you did from a review, so that it is clearly marked in this review issue? Were you able to verify the software for the remaining three checkboxes:

>  Functionality: Have the functional claims of the software been confirmed?
 Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
 Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
