### 2023-04-18 #4930 @kthyng
@earth-chris Just one comment on your paper: please update your .bib file with {} around words in which you need to preserve capitalization. For example, "California" and "Australia" are not capitalized — please check the rest of the references too.
