### 2023-04-18 9da35c9
updated capitalization, fixed broken/missing DOIs
