+++ docs/paper/paper.bib
@@ -1,7 +1,7 @@
 # journal articles
 
 @article{Grinnell:1917,
-  title={The niche-relationships of the California Thrasher},
+  title={The niche-relationships of the {California Thrasher}},
   author={Grinnell, Joseph},
   journal={The Auk},
   volume={34},
@@ -37,7 +37,7 @@
 }
 
 @article{Phillips:2017,
-  title={Opening the black box: An open-source release of Maxent},
+  title={Opening the black box: An open-source release of {Maxent}},
   author={Phillips, Steven J and Anderson, Robert P and Dud{\'\i}k, Miroslav and Schapire, Robert E and Blair, Mary E},
   journal={Ecography},
   volume={40},
@@ -49,7 +49,7 @@
 }
 
 @article{Phillips:2008,
-  title={Modeling of species distributions with Maxent: new extensions and a comprehensive evaluation},
+  title={Modeling of species distributions with {Maxent}: new extensions and a comprehensive evaluation},
   author={Phillips, Steven J and Dud{\'\i}k, Miroslav},
   journal={Ecography},
   volume={31},
@@ -61,7 +61,7 @@
 }
 
 @article{Booth:2014,
-  title={BIOCLIM: the first species distribution modelling package, its early applications and relevance to most current MAXENT studies},
+  title={{BIOCLIM}: the first species distribution modelling package, its early applications and relevance to most current {Maxent} studies},
   author={Booth, Trevor H and Nix, Henry A and Busby, John R and Hutchinson, Michael F},
   journal={Diversity and Distributions},
   volume={20},
@@ -73,7 +73,7 @@
 }
 
 @article{Nix:1986,
-  title={A biogeographic analysis of Australian elapid snakes},
+  title={A biogeographic analysis of {Australian} elapid snakes},
   author={Nix, Henry A},
   journal={Atlas of elapid snakes of Australia},
   volume={7},
@@ -82,7 +82,7 @@
 }
 
 @article{Merow:2013,
-  title={A practical guide to MaxEnt for modeling species' distributions: what it does, and why inputs and settings matter},
+  title={A practical guide to {Maxent} for modeling species' distributions: what it does, and why inputs and settings matter},
   author={Merow, Cory and Smith, Matthew J and Silander Jr, John A},
   journal={Ecography},
   volume={36},
@@ -94,7 +94,7 @@
 }
 
 @article{Elith:2011,
-  title={A statistical explanation of MaxEnt for ecologists},
+  title={A statistical explanation of {Maxent} for ecologists},
   author={Elith, Jane and Phillips, Steven J and Hastie, Trevor and Dud{\'\i}k, Miroslav and Chee, Yung En and Yates, Colin J},
   journal={Diversity and distributions},
   volume={17},
@@ -142,7 +142,7 @@
 }
 
 @article{Barber:2022,
-  title={Target-group backgrounds prove effective at correcting sampling bias in Maxent models},
+  title={Target-group backgrounds prove effective at correcting sampling bias in {Maxent} models},
   author={Barber, Robert A and Ball, Stuart G and Morris, Roger KA and Gilbert, Francis},
   journal={Diversity and Distributions},
   volume={28},
@@ -155,13 +155,13 @@
 
 @article{Kass:2021,
   author = {Kass, Jamie M. and Muscarella, Robert and Galante, Peter J. and Bohl, Corentin L. and Pinilla-Buitrago, Gonzalo E. and Boria, Robert A. and Soley-Guardia, Mariano and Anderson, Robert P.},
-  title = {ENMeval 2.0: Redesigned for customizable and reproducible modeling of species’ niches and distributions},
[…]
