+++ docs/paper/paper.bib
@@ -297,7 +297,7 @@
 
 @misc{gbif,
   author = {GBIF},
-  title = {GBIF: The Global Biodiversity Information Facility},
+  title = {{GBIF: The Global Biodiversity Information Facility}},
   year = {2022},
   publisher = {Global Biodiversity Information Facility},
   url = {https://www.gbif.org/what-is-gbif}
