+++ docs/paper/paper.bib
@@ -164,7 +164,7 @@
   doi = {10.1111/2041-210X.13628},
   url = {https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.13628},
   eprint = {https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/2041-210X.13628},
-  abstract = {Abstract Quantitative evaluations to optimize complexity have become standard for avoiding overfitting of ecological niche models (ENMs) that estimate species’ potential geographic distributions. ENMeval was the first R package to make such evaluations (often termed model tuning) widely accessible for the Maxent algorithm. It also provided multiple methods for partitioning occurrence data and reported various performance metrics. Requests by users, recent developments in the field, and needs for software compatibility led to a major redesign and expansion. We additionally conducted a literature review to investigate trends in ENMeval use (2015–2019). ENMeval 2.0 has a new object-oriented structure for adding other algorithms, enables customizing algorithmic settings and performance metrics, generates extensive metadata, implements a null-model approach to quantify significance and effect sizes, and includes features to increase the breadth of analyses and visualizations. In our literature review, we found insufficient reporting of model performance and parameterization, heavy reliance on model selection with AICc and low utilization of spatial cross-validation; we explain how ENMeval 2.0 can help address these issues. This redesigned and expanded version can promote progress in the field and improve the information available for decision-making.},
+  abstract = {Abstract Quantitative evaluations to optimize complexity have become standard for avoiding overfitting of ecological niche models (ENMs) that estimate species' potential geographic distributions. ENMeval was the first R package to make such evaluations (often termed model tuning) widely accessible for the Maxent algorithm. It also provided multiple methods for partitioning occurrence data and reported various performance metrics. Requests by users, recent developments in the field, and needs for software compatibility led to a major redesign and expansion. We additionally conducted a literature review to investigate trends in ENMeval use (2015–2019). ENMeval 2.0 has a new object-oriented structure for adding other algorithms, enables customizing algorithmic settings and performance metrics, generates extensive metadata, implements a null-model approach to quantify significance and effect sizes, and includes features to increase the breadth of analyses and visualizations. In our literature review, we found insufficient reporting of model performance and parameterization, heavy reliance on model selection with AICc and low utilization of spatial cross-validation; we explain how ENMeval 2.0 can help address these issues. This redesigned and expanded version can promote progress in the field and improve the information available for decision-making.},
   year = {2021}
 }
 
@@ -185,7 +185,7 @@
 
 @article{Ploton:2020,
   title={Spatial validation reveals poor predictive performance of large-scale ecological mapping models},
-  author={Ploton, Pierre and Mortier, Fr{\'e}d{\'e}ric and R{\'e}jou-M{\'e}chain, Maxime and Barbier, Nicolas and Picard, Nicolas and Rossi, Vivien and Dormann, Carsten and Cornu, Guillaume and Viennois, Ga{\"e}lle and Bayol, Nicolas and others},
+  author={Ploton, Pierre and Mortier, Fr{\'e}d{\'e}ric and R{\'e}jou-M{\'e}chain, Maxime and Barbier, Nicolas and Picard, Nicolas and Rossi, Vivien and Dormann, Carsten and Cornu, Guillaume and Viennois, Ga{\"e}lle and Bayol, Nicolas, and Lyapustin, Alexei and Gourlet-Fleury, Sylvie and P{\'e}lissier, Rapha{\"e}l},
   journal={Nature communications},
   volume={11},
   number={1},
@@ -212,7 +212,7 @@
                   Arnaud Joly and
                   Guillaume Lemaitre and
                   Jake Vanderplas and
-                  manoj kumar and
[…]
