### 2023-04-18 #4930 @earth-chris
@kthyng thanks for catching these! I've fixed the DOIs, updated capitalization where appropriate, and fixed some author name entries.
